% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSubTree.R
\name{buildSubTree}
\alias{buildSubTree}
\title{Generate nodes of a decision tree}
\usage{
buildSubTree(myData, classes, depth = 1, featureSpace = list(NA),
  random = T)
}
\arguments{
\item{myData}{Attribute data}

\item{classes}{Class data}

\item{depth}{The total number of layer for decision tree}

\item{random}{Do you want to the features data to be sample randomly when creating the tree?}
}
\value{
A list that contains a tree information
}
\description{
This is a recursive function which generates nodes of a decision tree. It is called in \code{get_Forest()} and in the function itself. It takes in several arguments, including the data and the current depth of the node. This function firstly calls \code{stopOrNot()}. If the returned value is true, then it means the terminal condition of building nodes is met, and it would create and return a node whose class is set to the majority class of data assigned to the node, using function \code{newNode()}. Next, \code{buildSubTree()} calls \code{bestRule()}, the returned object is of class ``Rule'', and it stores information of way to split the given data that can result in the minimum entropy. The returned rule is then stored in variable rulefound. Note that the function \code{bestRule()} sometimes return NA. This is because the data cannot be further split depend on the given data (for example, the features are all categorical, and features of the given data are identical). In this case a terminal node is returned. If the terminal condition if not met, and if there is a usable rule to split the data, then the data is split into two groups subData1 and subData2 according to the best rule. Each of the sub-groups is then passed as argument to \code{buildSubTree()} in order to obtain child nodes. At last, a node is built using \code{newNode()} with rulefound as split rule and returned objects of \code{buildSubTree()} as child nodes.
}
