% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_graph.R
\name{dt_graph}
\alias{dt_graph}
\title{Plot decision tree}
\usage{
dt_graph(or_list, shape = "rectangle", arrow_size = 0.6, circular = F,
  node_fit = 0)
}
\arguments{
\item{or_list}{A decision tree list}

\item{shape}{Node shape}

\item{arrow_size}{Size of pointing arrow}

\item{circular}{do you want the node shape to be circular?}

\item{node_fit}{Size of node frame}
}
\value{
A tree-like graph
}
\description{
This function is used to plot decision tree graph given the output of \code{buildSubTree()}. The arguments of this function include the input list from \code{buildSubTree()}, node shape, arrow size, node size fit and whether it is a circular tree graph. In order to use the plotting function in package igraph, the judgments and directions of each nodes as well as edge directions are needed. Since the format of the input is lists nested by lists as the tree grows, this function first unlists and create a new vector including judgements, classifications and their positions. Then, the directions of each node are created from the vector above. 'x' represents left and 'y' represents right. After using node directions to create edge directions, everything that \code{plot()} function in igraph package requires is created. The final step of this function is to plot the decision tree based on the results above. In addition, some arguments, like arrow size and node shape, are added to improve the basic tree graph.
}
