% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_node.R
\name{find_node}
\alias{find_node}
\title{Find the best node with minimum entropy}
\usage{
find_node(features, class)
}
\arguments{
\item{features}{Attribute data}

\item{class}{class data}
}
\value{
A list contains the best node position and its entropy
}
\description{
This function is to determine the best node to split data. The criteria of best node is to have minimum split entropy since less entropy indicates more homogeneity in the group. Among all the possible splitting results, \code{get_split_entropy} is called within a loop in this function and returned the best node as well as the corresponding minimum entropy.
}
