% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Calculate the single entropy}
\usage{
get_entropy(features, class)
}
\arguments{
\item{features}{Attribute data}

\item{class}{class data}
}
\value{
A numeric entropy value
}
\description{
To calculate entropy for single branch, the data is first separated into two sections. One is feature section which contains the attributes and the other one is class section which contains the classification result in the dataset. The two inputs in this function feature and class are corresponded to these sections. For each unique class existed in the class section, $-p_ilog_2p_i$ term is calculated and summed for all classes to obtain total initial single entropy.
}
