% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Forest.R
\name{get_Forest}
\alias{get_Forest}
\title{Generate a forest}
\usage{
get_Forest(feature, class, NumofBag, NumofObs)
}
\arguments{
\item{class}{class data}

\item{NumofBag}{The number of tree in the forest}

\item{NumofObs}{The number of observation in a single tree}
}
\value{
A list of multiple tree lists.
}
\description{
This function is used to generate multiple decision tree. A technique called Bagging is applied to reduce variance and increase the prediction accuracy (James, Witten, Hastie, \&Tibshirani, 2017). Specifically, the number of features used in bootstrapped training set is determined to be
\deqn{m=\sqrt{p}} This function takes four inputs: feature, class, NumofBag and NumofObs. NumofBag is the number of bootstrapped training set needed to generate and NumofObs the the number of observation per training set. Given these two inputs, bootstrapped features set and class set are sampled from original features section and class section. With the generated bootstrapped set, \code{buildSubTree()} is called to generate the single decision tree. Accordingly, an aggregated forest list contained multiple tree list is returned.
}
