% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split_entropy.R
\name{get_split_entropy}
\alias{get_split_entropy}
\title{Calculate spilit entropy}
\usage{
get_split_entropy(features, class, node)
}
\arguments{
\item{features}{Attribute data}

\item{class}{class data}

\item{node}{The node at which you want to calculate the split entropy}
}
\value{
A numeric split entropy value
}
\description{
For this function, there are total three inputs: features, class and node where features and class are the same with previous inputs while node stands for the ``position'' used to split the attribute and class correspondingly.
\code{get_entropy} is called to calculate the single entropy for both group based on the node value. The total split entropy value would thus be calculated and returned based on the formula \deqn{E(T,X)=\sumP(A)E(A)}
where X is our groups after splitting. Since we only considered binary splitting in our project, so we always divided the data into two groups. .
}
