% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRule.R
\name{newRule}
\alias{newRule}
\title{Creat a new Rule class}
\usage{
newRule(feature, judgement, value)
}
\arguments{
\item{feature}{The name of the variable to be split}

\item{judgement}{Comparison operator}

\item{value}{The number that is used to compare}
}
\value{
An object of class ''Rule''
}
\description{
This function return an object of class ''Rule''. It takes in three arguments: feature, judgement and value. Feature stores a name of the variable to be split, judgment states how to split the data according the value. In this case, judgment is always set to the character ``($<$)=``. When the variable is categorical, judgement should be seen as ``=``, otherwise it is viewed as ``$<$=``. Lastly, value stores the categorical or numeric value of the rule. In this function, the three arguments are stored in a single list, and the list is of class ``Rule''.
}
