% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTree.R
\name{predictTree}
\alias{predictTree}
\title{Predict the class of new data}
\usage{
predictTree(myData, Tree)
}
\arguments{
\item{myData}{New test data}

\item{Tree}{The decision tree model}
}
\value{
A character class
}
\description{
This function is used in \code{randforePredict()}, it is designed to use decision tree models to predict the class of new data. Its inputs is a single-row new data and an object of decision tree model, and its output is character class. This function has a recursive structure. It firstly set the tree as the current node, and then according to the splitting rule and data, the current node is set either to childT(when test data satisfy the splitting rule) or childF otherwise. The split rule would be NA at the leaves, the class of the corresponding node would be returned.
}
