% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randforePredict.R
\name{randforePredict}
\alias{randforePredict}
\title{Make prediction with random forest}
\usage{
randforePredict(ForestList, myData)
}
\arguments{
\item{ForestList}{A List of multiple decision trees (Forest model)}

\item{myData}{Test data}
}
\value{
A vector of predicted class
}
\description{
This function is used to make predictions based on the random forest model. It takes two inputs: list of built decision trees returned by \code{get\_Forest()} and test data as myData. This function returns a vector of predicted classes. For each row of the data given, a list of predictions is initialized and filled using function \code{predictTree()}. Each cell of the prediction list is then filled using the corresponding decision tree in the ForestList and data in the corresponding row. Since random forest usually takes majority vote for classification problems, so the most occurring class in the prediction list would be chosen as the predicted class for that row of data.
}
